# apps/golearnix-auth/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/api/v1/auth/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Borra al usuario autenticado y todas sus sesiones.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "User"
        ],
        "summary": "Eliminar cuenta de usuario",
        "responses": {
          "200": {
            "description": "Usuario eliminado exitosamente",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "401": {
            "description": "No autorizado",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Valida credenciales y devuelve un JWT válido por 24h.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar sesión de usuario",
        "parameters": [
          {
            "description": "Email y contraseña",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Credenciales inválidas",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "401": {
            "description": "Credenciales incorrectas",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Invalida el JWT actual eliminando la sesión en BD.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Cerrar sesión",
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "401": {
            "description": "Token no proporcionado o inválido",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Error interno al cerrar sesión",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Devuelve los datos del usuario autenticado.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "User"
        ],
        "summary": "Obtener información del usuario actual",
        "responses": {
          "200": {
            "description": "Información del usuario",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "No autorizado",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Crea un usuario con nombre, email, contraseña y rol.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Registrar un nuevo usuario",
        "parameters": [
          {
            "description": "Datos de registro",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Datos inválidos",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "409": {
            "description": "Email ya registrado",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/validate": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Comprueba que el token está activo y devuelve email y tiempo restante.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Validar token JWT",
        "responses": {
          "200": {
            "description": "Token válido",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "Token inválido o expirado",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "request.LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "request.RegisterRequest": {
      "type": "object",
      "required": [
        "email",
        "name",
        "password",
        "role"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 2
        },
        "password": {
          "type": "string",
          "minLength": 8
        },
        "role": {
          "type": "string",
          "enum": [
            "admin",
            "instructor",
            "student"
          ]
        }
      }
    },
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Token JWT en formato Bearer. Ejemplo: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}